BXB-4263 Enable CMake for jsoncpp 1.9.5
A change in jsoncpp's CMake file led to it naming the library targets
based off of the PROJECT_NAME variable. So this was incorrectly adding
the jsoncpp library target as "mediacore_lib" or something similar
depending on the top level directory. Manually setting this to jsoncpp
makes the library target created as expected. Additionally, a change led
to BUILD_SHARED_LIBS needing to be turned on manually.